Major rework of type inference
The constraints are now stored in union-find sets instead of
"linear" links which simplifies the type inferencing. It makes
it easier to figure out the correct type in both "directions".

Previously the types needed to be solved in a linear way from top
to bottom according to how the variables/values where used. Now
all uses for different variables/values are stored in sets which
means that types can be solved order independant. fixes #3