Implement logic to check that functions called on generics are valid
If a struct has a generic type `T`, every function call on that
generic type needs to have a corresponding function in a trait
that is enforced on `T`. Logic have been implemented which ensures
that every method call on a variable with type `T`, and every
static function call on type `T`, is valid inside the structs
impl block.