Add primite check to ensure func call generics are correct
Since there are currently no way to infer the type of generics for
functions if they are not specified at the call site, this check
is added to make sure that generics always are specified.
This will not be the case in the future, so this check will only
exist temporarily and will be removed when it isn't needed anymore.