Fix issue with types when parsing potential generic list
When unsure if a less than sign is a "lt" or start of a generic
list, new types are created for the items in the potential generic
list even though they might not actually be types. These were then
never removed when it turned out that it was a "lt" symbol and the
parsed tokens doesn't represent types.

The fix is to remove the types if it turns out that the less than
symbol wasn't a start of a generic list.