Make inc/dec expression instead of statements
The increment and decrement operator have been changed. Previously
they were specifed at the start of a expression and wrapped the
expression and made it a statement ex. "++x". This has no been
changed so that one is allowed to use inc/dec in nested expression.

The new syntax is ".++" and ".--" which follows the syntax of most
unary operators following the expression that should be updated.
The inc/dec is ALWAYS evaluated before its use, so it is similar to
a prefix inc/dec in other languages.