Fix problem when var and fn have same name
There is a possiblity that a variable containing a function pointer
and a "normal" function have the same name. In those cases, there
can be clashes where the compiler doesn't know which is which.
This commit fixes one issue when calling a function with the same
name as a variable in the current scope.