Update how program handles `TypeId`s & `Ty`s
The TyEnv (type environment) have been moved to a global static
variable which makes it accessable from the whole program.
This is needed since a lot of types/structs contains `TypeId`s
which need to be converted to their responding `Ty`s during for
example compares and hashing.

Before this commit, some compares and hashes had to be done on the
`TypeId`s directly since the TyEnv wasn't accessable from anywhere
which could cause miss-compilations.