Change format of method names in preparation for wasm
The name of a method belonging to a specific ADT was formatted as:
  <ADT_NAME>-<METHOD_NAME>
The separator is now a dot instead:
  <ADT_NAME>.<METHOD_NAME>
This is done since a dot is more "familiar" looking at it from a
human perspective. And with the focus on wasm support, these
methods will very likely be called from Javascript code which means
that a person will read/write them (which wasn't the case before).