Implement `Eq<T>` trait for equality comparisong of ADT
This new trait now allows one to use the `==` & `!=` symbols to
compare ADTs. The compiler will, if the operands to the equality
are ADTs, rewrite the `==` & `!=` operators into `eq` calls on the
`Eq<T>` trait for the ADTs.

If the compared ADTs doesn't implement the `Eq<T>` trait, a compile
error will occure.