Implement `Cmp<T>` trait for comparison of ADT
This new trait now allows one to use the `>`, `<`, `>=` & `<=`
symbols to compare ADTs. The compiler will, if the operands to the
comparison are ADTs, rewrite the `>`, `<`, `>=` & `<=` operators
into `cmp` calls on the `Cmp<T>` trait for the ADTs.

If the compared ADTs doesn't implement the `Cmp<T>` trait, a compile
error will occure.